// types/access-incoming.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface AccessIncoming {
  /**
   * The type of resource request as a string.  This field defines which other fields are allowed in the request object.
   */
  type: "incoming-payment";
  /**
   * The types of actions the client instance will take at the RS as an array of strings.
   */
  actions: ("create" | "complete" | "read" | "read-all" | "list" | "list-all")[];
  /**
   * A string identifier indicating a specific resource at the RS.
   */
  identifier?: string;
  [k: string]: unknown;
}