// types/access-quote.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface AccessQuote {
  /**
   * The type of resource request as a string.
   */
  type: "quote";
  /**
   * The types of actions the client instance will take at the RS.
   */
  actions: ("create" | "read" | "read-all")[];
  [k: string]: unknown;
}