// types/continue.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * The AS determines that the request can be continued.
 */
export interface Continue {
  /**
   * Continuation access token.
   */
  access_token: {
    value: string;
    [k: string]: unknown;
  };
  /**
   * The URI for continuation requests.
   */
  uri: string;
  /**
   * The interaction reference from the sender's confirmation.
   */
  interact_ref?: string;
  /**
   * The wait time before continuing.
   */
  wait?: number;
  [k: string]: unknown;
}