// types/manage-token.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ManageAccessToken {
  /**
   * The action we want do on the token
   */
  action?: "rotate" | "revoke";
  /**
   * The token management URL
   */
  url: string;
  /**
   * The previous token that should be rotated
   */
  access_token: string;
  [k: string]: unknown;
}