// types/wallet-address.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * The assetCode is a code that indicates the underlying asset. An ISO4217 currency code should be used whenever possible. The ISO4217 representation of the US Dollar is USD.
 */
export type AssetCode = string;
/**
 * The number of decimal places that defines the scale of the smallest divisible unit for the given asset code. It determines how an integer amount is scaled to derive the actual monetary value. For example, USD has an asset scale of 2 with the smallest unit being 0.01. An integer amount of `1000` with an `assetCode` of `USD` and `assetScale` of `2` translates to $10.00.
 */
export type AssetScale = number;

/**
 * A **wallet address** resource is the root of the API and contains the public details of the financial account represented by the Wallet Address that is also the service endpoint URL.
 */
export interface WalletAddress {
  /**
   * The URL identifying the wallet address.
   */
  id: string;
  /**
   * A public name for the account. This should be set by the account holder with their provider to provide a hint to counterparties as to the identity of the account holder.
   */
  publicName?: string;
  assetCode?: AssetCode;
  assetScale?: AssetScale;
  /**
   * The URL of the authorization server endpoint for getting grants and access tokens for this wallet address.
   */
  authServer?: string;
  /**
   * The URL of the resource server endpoint for performing Open Payments with this wallet address.
   */
  resourceServer?: string;
  [k: string]: unknown;
}